- Add screen ProductListFragment.kt and its view model ProductListViewModel.kt - Add recycler view in ProductListFragment.kt to show list of products - Add recycler view adapter ProductRecyclerViewAdapter.kt which inherits ListAdapter and compare old list and new list using DiffUtil.ItemCallback - Depend on Picasso lib to load images - Add TextInputLayout in ProductListFragment.kt to search for a product - Apply dependency injection on the ui layer using Hilt dagger

+++ app/src/main/java/com/coolblue/miniapp/model/entities/Product.kt
@@ -6,7 +6,7 @@ data class Product(
     val reviewInformation: ReviewInformation,
     val uSPs: List<String>,
     val availabilityState: Int,
-    val salesPriceIncVat: Int,
+    val salesPriceIncVat: Double,
     val productImage: String,
     val coolbluesChoiceInformationTitle: String,
     val promoIcon: PromoIcon,
